LayoutTests/imported/w3c/web-platform-tests/shadow-dom/Document-prototype-currentScript-expected.txt

CONSOLE MESSAGE: line 2422: Error: assert_equals: expected null but got Element node <script id="outerScriptElement">

var outerScriptElement ...
CONSOLE MESSAGE: line 2422: Error: assert_equals: expected null but got Element node <script id="outerScriptElement">

var outerScriptElement ...

Harness Error (FAIL), message = Error: assert_equals: expected null but got Element node <script id="outerScriptElement">

var outerScriptElement ...

PASS document.currentScript must not to be set to a script element in a shadow tree in open mode 
PASS document.currentScript must not to be set to a script element in a shadow tree in closed mode 
PASS document.currentScript must be set to a script element that loads an external script in a document tree 
PASS document.currentScript must be set to a script element that loads an external script in a document tree 
PASS document.currentScript must not be set to a script element that loads an external script in an open shadow tree 
PASS document.currentScript must not be set to a script element that loads an external script in a closed shadow tree 
PASS document.currentScript must be set to a script element that loads an external script that was in an open shadow tree and then removed 
PASS document.currentScript must be set to a script element that loads an external script that was in a closed shadow tree and then removed 
